# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestion de Cobros - API
  version: 1.0.0
  description: |
    API REST del Sistema de Gestion de Cobros (recaudacion, conciliacion bancaria,
    importacion de extractos, gastos, reportes y auditoria).

    Base path: `/api`. Autenticacion mediante JWT de acceso enviado como
    `Authorization: Bearer <accessToken>`.

    Formato de listas: `{ "data": [...], "meta": { page, pageSize, total, totalPages } }`.
    Formato de errores: `{ "error": { "code", "message", "details?" } }`.

    Los montos y tasas viajan como STRING en JSON (p. ej. `"3600.00"`) para
    preservar la precision de `DECIMAL(18,2)` / `DECIMAL(18,6)`.
servers:
  - url: 'http://localhost:{port}/api'
    description: Servidor local de desarrollo
    variables:
      port:
        default: '4000'
        description: Puerto del API (variable de entorno PORT)
security:
  - bearerAuth: []
...
paths:
...
  /pagos:
    post:
      tags: [Pagos]
      operationId: reportarPago
      summary: Reportar un pago
      description: |
        Requiere permiso: pagos.reportar. La tasa se deriva de montoBs / montoUsd
        (nunca se acepta como entrada). `cobradorId` solo lo pueden usar usuarios
        con permiso pagos.ver_todos para reportar a nombre de un cobrador.
        `bancoOrigenId` es opcional: si el parametro `pago.banco_origen_obligatorio`
        esta activo ('1', valor por defecto) y no se envia, el API responde 400 con
        "El banco de origen es obligatorio.".

        Los montos `montoBs` y `montoUsd` deben resolver a un decimal finito y
        estrictamente positivo. Un valor no numerico, cero o negativo responde **400**
        con "El monto debe ser un numero positivo" (regla compartida con
        `PUT /pagos/{id}`), evitando que el error derive en un 500.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportarPagoRequest'
            example:
              fechaPago: '2026-09-19'
              fechaDocumento: '2026-08-01'
              referencia: '12345678'
              bancoOrigenId: 1
              cuentaRecaudadoraId: 1
              montoBs: '3600.00'
              montoUsd: '20.00'
              cliente: Juan Perez
              concepto: Pago cuota mensual
              tipoCobro: nuevo
              tipoPagoId: 1
              observaciones: Reportado en ruta
              cobradorId: 1
      responses:
        '201':
          description: Pago reportado (estado pendiente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Pagos]
      operationId: listarPagos
      summary: Listar pagos reportados
      description: |
        Requiere permiso: pagos.ver_todos o pagos.ver_propios. El aislamiento por
        cobrador se aplica en la capa de datos cuando el usuario no tiene pagos.ver_todos.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - name: fechaDesde
          in: query
          required: false
          schema: { type: string, format: date, example: '2026-09-01' }
        - name: fechaHasta
          in: query
          required: false
          schema: { type: string, format: date, example: '2026-09-30' }
        - name: cobradorId
          in: query
          required: false
          schema: { type: integer, example: 1 }
        - name: bancoOrigenId
          in: query
          required: false
          schema: { type: integer, example: 1 }
        - name: cuentaRecaudadoraId
          in: query
          required: false
          schema: { type: integer, example: 1 }
        - name: estado
          in: query
          required: false
          schema: { type: string, enum: [pendiente, validado, rechazado, duplicado], example: pendiente }
        - name: tipoCobro
          in: query
          required: false
          schema: { type: string, enum: [nuevo, viejo], example: nuevo }
        - name: tipoPagoId
          in: query
          required: false
          description: Filtra por tipo de pago
          schema: { type: integer, example: 1 }
        - name: referencia
          in: query
          required: false
          schema: { type: string, maxLength: 60, example: '12345678' }
        - name: montoMin
          in: query
          required: false
          schema: { type: number, example: 10 }
        - name: montoMax
          in: query
          required: false
          schema: { type: number, example: 100 }
        - name: soloRevisar
          in: query
          required: false
          description: Filtra pagos con clasificacion a revisar
          schema: { type: string, enum: ['true', 'false'], example: 'true' }
      responses:
        '200':
          description: Lista paginada de pagos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagoPage'
              example:
                data:
                  - id: 10
                    fechaPago: '2026-09-19T00:00:00.000Z'
                    referencia: '12345678'
                    montoBs: '3600.00'
                    montoUsd: '20.00'
                    tasa: '180.000000'
                    cliente: Juan Perez
                    concepto: Pago cuota mensual
                    tipoPagoId: 1
                    soporteUrl: /uploads/1699999999999-comprobante.pdf
                    tipoCobro: nuevo
                    tipoCobroDerivado: nuevo
                    revisarClasificacion: false
                    observaciones: null
                    estado: pendiente
                    motivoRechazo: null
                    validadoAt: null
                    createdAt: '2026-09-19T14:00:00.000Z'
                    cobradorId: 1
                    bancoOrigenId: 1
                    cuentaRecaudadoraId: 1
                    movimientoBancoId: null
                    cobrador: { id: 1, nombre: Pedro Rojas, codigo: CBR-001 }
                    tipoPago: { id: 1, nombre: Pago Movil }
                    bancoOrigen: { id: 1, nombre: Banco de Venezuela, codigo: BDV }
                    cuentaRecaudadora: { id: 1, numeroCuenta: '0102-0123-45-6789012345', alias: Cuenta principal, banco: { id: 1, nombre: Banco de Venezuela } }
                    movimientoBanco: null
                    validador: null
                meta: { page: 1, pageSize: 20, total: 1, totalPages: 1 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/validar-lote:
    post:
      tags: [Conciliacion]
      operationId: validarLote
      summary: Validar un lote de pagos
      description: |
        Requiere permiso: pagos.validar_lote. Valida pares pago/movimiento en una
        transaccion. Los items fallidos se devuelven en `errores`, el resto se procesa.
        CR-001: si un pago coincide por referencia + monto en Bs con un movimiento YA
        conciliado con otro pago (posible duplicado, misma cuenta recaudadora y sin
        ventana de fecha), NO se valida y se reporta en `errores` con un motivo que
        contiene "posible duplicado"; el pago queda `pendiente` para validacion manual.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [pagoReportadoId, movimientoBancoId]
                    properties:
                      pagoReportadoId: { type: integer, example: 10 }
                      movimientoBancoId: { type: integer, example: 25 }
            example:
              items:
                - { pagoReportadoId: 10, movimientoBancoId: 25 }
                - { pagoReportadoId: 11, movimientoBancoId: 26 }
      responses:
        '200':
          description: Resultado de la validacion por lote
          content:
            application/json:
              schema:
                type: object
                properties:
                  procesados: { type: integer, example: 1 }
                  errores:
                    type: array
                    items:
                      type: object
                      properties:
                        pagoReportadoId: { type: integer, example: 11 }
                        motivo: { type: string, example: Movimiento no disponible }
              example:
                procesados: 1
                errores:
                  - pagoReportadoId: 11
                    motivo: El pago coincide con un movimiento ya conciliado con otro pago (posible duplicado)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Pagos]
      operationId: obtenerPago
      summary: Obtener pago por id
      description: 'Requiere permiso: pagos.ver_todos o pagos.ver_propios (con aislamiento por cobrador).'
      responses:
        '200':
          description: Pago encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Pagos]
      operationId: editarPago
      summary: Editar pago reportado (incluye pagos validados)
      description: |
        Permisos: la ruta exige `pagos.reportar` **o** `pagos.editar`; sin ninguno
        de los dos responde **403** en el guard. Un pago en estado `pendiente` se
        edita con cualquiera de esos dos permisos (el cobrador sigue editando los
        propios). Editar un pago `validado` exige ADEMAS el permiso elevado
        `pagos.editar`; con solo `pagos.reportar` responde **403** desde el servicio.
        `pagos.ver_todos` es un permiso de **solo lectura** y NO habilita la edicion.
        Los pagos `rechazado` y `duplicado` no son editables: responden **409**. El
        permiso elevado se verifica en el servicio contra el estado leido dentro de
        la transaccion, de modo que la edicion del cobrador sobre sus propios pagos
        pendientes no se rompe.

        Edicion de un pago `validado` (en caliente, dentro de una transaccion): se
        reevalua el vinculo con el movimiento bancario usando los valores NUEVOS
        (referencia, monto dentro de `match.amount_tolerance_bs`, fecha dentro de la
        ventana de `match.date_window_days` y misma cuenta recaudadora). Si el
        vinculo se mantiene, se recalcula `conciliaciones.diferenciaBs` y se registra
        la auditoria (`datosAntes`/`datosDespues`) en la MISMA transaccion; el pago
        conserva `estado = 'validado'`. Si la edicion rompe el vinculo, responde
        **409** con el movimiento en conflicto y no modifica nada: primero hay que
        revertir la validacion (`POST /pagos/{id}/revertir`). Nuevos endpoints: 0.

        Si cambian los montos, se recalcula la tasa; ambos montos deben resolver a un
        decimal finito y estrictamente positivo (`El monto debe ser un numero positivo`,
        **400**), la misma regla que aplica `POST /pagos`. Actualizacion parcial: un
        `bancoOrigenId` ausente deja el banco guardado intacto, mientras que `null`
        lo limpia. La regla de obligatoriedad se evalua sobre el valor resultante:
        con `pago.banco_origen_obligatorio` activo (`'1'`) un pago sin banco
        responde 400 "El banco de origen es obligatorio.".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                referencia: { type: string, maxLength: 60, example: '87654321' }
                bancoOrigenId:
                  type: integer
                  nullable: true
                  description: >-
                    Banco de origen opcional. Ausente deja el banco actual sin cambios;
                    null lo limpia. Si pago.banco_origen_obligatorio esta activo ('1') y
                    el valor resultante queda nulo, responde 400.
                  example: 1
                cuentaRecaudadoraId: { type: integer, example: 1 }
                montoBs: { type: string, example: '4000.00' }
                montoUsd: { type: string, example: '22.00' }
                fechaPago: { type: string, format: date, example: '2026-09-19' }
                cliente: { type: string, maxLength: 180, example: Juan Perez }
                concepto: { type: string, maxLength: 255, example: Pago cuota mensual }
                tipoCobro: { type: string, enum: [nuevo, viejo], example: nuevo }
                tipoPagoId:
                  type: integer
                  description: >-
                    Tipo de pago opcional. Debe OMITIRSE para no modificar el valor actual;
                    enviar null responde 400 (a diferencia de bancoOrigenId, que si acepta null
                    para limpiarlo).
                  example: 1
                observaciones: { type: string, maxLength: 2000, example: Corregido por el cobrador }
              example:
                referencia: '87654321'
                montoBs: '4000.00'
                montoUsd: '22.00'
      responses:
        '200':
          description: Pago actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >-
            El usuario no tiene `pagos.reportar` ni `pagos.editar` (guard de la ruta),
            o el pago esta `validado` y carece del permiso elevado `pagos.editar`
            (verificado en el servicio contra el estado fresco).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: forbidden
                  message: 'No tiene permiso para editar un pago validado. Se requiere el permiso "pagos.editar".'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            El pago esta `rechazado`/`duplicado` (no editable) o la edicion rompe el
            vinculo con el movimiento bancario vinculado. En ambos casos no se
            modifica nada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                estadoNoEditable:
                  value:
                    error:
                      code: conflict
                      message: 'No se puede editar un pago en estado "rechazado"'
                vinculoRoto:
                  value:
                    error:
                      code: conflict
                      message: 'La edicion rompe la conciliacion con el movimiento bancario #42 (referencia 120012345678, Bs 4000.00). Revierta la validacion antes de editar.'
                      details: { movimientoBancoId: 42, referencia: '120012345678' }
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Pagos]
      operationId: eliminarPago
      summary: Eliminar pago reportado no validado
      description: |
        Requiere permiso: pagos.eliminar. Es un borrado FISICO (hard delete) y solo se
        permite mientras el pago NO este `validado` (es decir, en estado `pendiente`,
        `rechazado` o `duplicado`). Un pago no validado no tiene fila en `conciliaciones`
        ni `movimientoBancoId`, por lo que no deja referencias huerfanas.
        Responde 409 si el pago esta `validado`: primero debe revertirse la validacion
        (`POST /pagos/{id}/revertir`). El aislamiento por cobrador aplica: un usuario sin
        pagos.ver_todos solo alcanza sus propios pagos.
        La instantanea completa del registro se escribe en la auditoria (entidad
        `pagos_reportados`, accion `borrar`, con `datosAntes`) dentro de la MISMA
        transaccion que el borrado, de modo que ambos succeed o fallan juntos.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Pago eliminado (sin contenido)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /pagos/{id}/coincidencias:
    get:
      tags: [Conciliacion]
      operationId: obtenerCoincidencias
      summary: Buscar coincidencias bancarias
      description: |
        Requiere permiso: pagos.validar. Devuelve los movimientos bancarios
        candidatos puntuados (0..100) para el pago indicado. `data` conserva la
        forma original (array de coincidencias). `duplicado` es una senal ADITIVA y
        nullable: cuando NO es `null`, el pago coincide por referencia + monto en Bs
        con un movimiento bancario YA conciliado con OTRO pago (misma cuenta
        recaudadora, sin ventana de fecha), por lo que la validacion automatica no
        procede (ver `POST /pagos/{id}/validar`).
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Coincidencias encontradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Coincidencia' }
                  duplicado:
                    description: >-
                      Senal de duplicado (aditiva): un movimiento ya conciliado con
                      otro pago que coincide por referencia + monto en Bs, o `null`
                      si el pago no es duplicado.
                    nullable: true
                    allOf:
                      - $ref: '#/components/schemas/DuplicadoInfo'
              example:
                data:
                  - movimiento: { id: 25, referencia: '12345678', montoBs: '3600.00', fechaEjecucion: '2026-09-19T00:00:00.000Z' }
                    puntaje: 90
                    referenciaExacta: true
                    coincidenciaSufijo: false
                    diferenciaMontoBs: '0.00'
                    diferenciaDias: 0
                duplicado: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/{id}/validar:
    post:
      tags: [Conciliacion]
      operationId: validarPago
      summary: Validar pago contra un movimiento
      description: |
        Requiere permiso: pagos.validar. Si se omite `movimientoBancoId`, se usa la
        mejor coincidencia automatica: el contraste de referencia usa el ultimo
        `match.reference_suffix` (8) de la referencia reportada, o sus propios digitos
        si es mas corta, con un minimo de 4 digitos (una referencia con menos de 4
        digitos ya no matchea por sufijo, solo por igualdad exacta); exige ademas el
        monto en Bs dentro de `match.amount_tolerance_bs` y la fecha dentro de
        `match.date_window_days`. Marca el pago como `validado` y el movimiento como
        `conciliado`.
        CR-001: si el pago coincide con un movimiento YA conciliado con otro pago
        (posible duplicado: misma cuenta recaudadora y referencia + monto en Bs, SIN
        ventana de fecha), la via automatica responde **409** y NO valida: el pago
        permanece `pendiente` y debe resolverse manualmente enviando el
        `movimientoBancoId`. La via manual (con `movimientoBancoId`) no se ve afectada.
        Ejemplo de request: `POST /api/pagos/10/validar`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                movimientoBancoId:
                  type: integer
                  description: Movimiento bancario a conciliar (opcional)
                  example: 25
            example:
              movimientoBancoId: 25
      responses:
        '200':
          description: Pago validado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            Validacion automatica (sin `movimientoBancoId`) y el pago coincide con un
            movimiento ya conciliado con otro pago (posible duplicado): no se valida y
            el pago queda `pendiente` para validacion manual. Tambien responde 409 si
            el pago o el movimiento ya fueron procesados.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: conflict
                  message: >-
                    El pago coincide con un movimiento bancario ya conciliado con otro
                    pago. No se puede validar automaticamente: seleccione manualmente
                    el movimiento para validarlo.
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/{id}/rechazar:
    post:
      tags: [Conciliacion]
      operationId: rechazarPago
      summary: Rechazar pago reportado
      description: 'Requiere permiso: pagos.rechazar. El motivo es obligatorio (minimo 3 caracteres).'
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [motivoRechazo]
              properties:
                motivoRechazo: { type: string, minLength: 3, maxLength: 1000, example: Referencia no encontrada en el extracto }
            example:
              motivoRechazo: Referencia no encontrada en el extracto
      responses:
        '200':
          description: Pago rechazado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/{id}/duplicado:
    post:
      tags: [Conciliacion]
      operationId: marcarDuplicado
      summary: Marcar pago como duplicado
      description: 'Requiere permiso: pagos.marcar_duplicado.'
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Pago marcado como duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /pagos/{id}/revertir:
    post:
      tags: [Pagos]
      operationId: revertirPago
      summary: Revertir la validacion de un pago
      description: |
        Requiere permiso: pagos.revertir_validacion. Es la operacion INVERSA exacta de
        `POST /pagos/{id}/validar` y solo se permite mientras el pago esta `validado`
        (409 si no lo esta). Dentro de una sola transaccion: (a) elimina la fila de
        `conciliaciones`, (b) devuelve el movimiento bancario enlazado a `no_conciliado`
        para que pueda conciliarse de nuevo, y (c) reinicia el pago al estado destino,
        limpiando `movimientoBancoId`, `validadoPor` y `validadoAt` (y fijando o
        limpiando `motivoRechazo` segun corresponda). Se registran en auditoria las
        instantaneas anteriores y posteriores.
        El cuerpo exige `estado` (`pendiente` o `rechazado`); `motivoRechazo` es
        obligatorio (minimo 3, maximo 1000 caracteres) cuando `estado` es `rechazado`
        y no debe enviarse cuando `estado` es `pendiente`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevertirPagoRequest'
            example:
              estado: rechazado
              motivoRechazo: Validacion revertida por error de conciliacion
      responses:
        '200':
          description: Pago revertido; devuelve el pago actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pago'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token JWT de acceso (accessToken) enviado como Authorization: Bearer <token>.'
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: Numero de pagina (>= 1). Por defecto 1.
      schema: { type: integer, minimum: 1, example: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      description: Tamano de pagina (1..200). Por defecto 20.
      schema: { type: integer, minimum: 1, maximum: 200, example: 20 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, example: 1 }
...
  responses:
    BadRequest:
      description: Solicitud invalida (validacion Zod o regla de negocio)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_error
              message: Datos invalidos
              details: {}
    Unauthorized:
      description: No autenticado (token ausente, invalido o expirado)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: unauthorized, message: Token de acceso requerido }
    Forbidden:
      description: Permisos insuficientes
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: forbidden, message: 'Permisos insuficientes: pagos.validar' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: not_found, message: Recurso no encontrado }
    Conflict:
      description: Conflicto (duplicado o estado incompatible)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: conflict, message: El registro ya existe (valor duplicado) }
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: internal_error, message: Error interno del servidor }
  schemas:
...
    ReportarPagoRequest:
      type: object
      required: [fechaPago, referencia, cuentaRecaudadoraId, montoBs, montoUsd, tipoCobro]
      properties:
        fechaPago:
          type: string
          description: Fecha de pago (fecha de calendario)
          example: '2026-09-19'
        fechaDocumento:
          type: string
          description: Fecha del documento/usado para clasificar antiguedad (opcional)
          example: '2026-08-01'
        referencia: { type: string, maxLength: 60, example: '12345678' }
        bancoOrigenId:
          type: integer
          nullable: true
          description: >-
            Banco de origen opcional. Si el parametro pago.banco_origen_obligatorio
            esta activo ('1', valor por defecto) y no se envia, el API responde 400 con
            "El banco de origen es obligatorio.".
          example: 1
        cuentaRecaudadoraId: { type: integer, example: 1 }
        montoBs:
          type: string
          description: >-
            Monto en bolivares como string (DECIMAL(18,2)). Debe ser estrictamente
            positivo (400 "El monto debe ser un numero positivo" si es no numerico,
            cero o negativo).
          example: '3600.00'
        montoUsd:
          type: string
          description: >-
            Monto en dolares como string (DECIMAL(18,2)). Debe ser estrictamente
            positivo (400 "El monto debe ser un numero positivo" si es no numerico,
            cero o negativo).
          example: '20.00'
        cliente: { type: string, maxLength: 180, example: Juan Perez }
        concepto: { type: string, maxLength: 255, example: Pago cuota mensual }
        tipoCobro: { type: string, enum: [nuevo, viejo], example: nuevo }
        tipoPagoId:
          type: integer
          description: >-
            Referencia opcional al catalogo tipos_pago (debe existir y estar activo). Debe
            OMITIRSE cuando no aplica; enviar null responde 400.
          example: 1
        observaciones: { type: string, maxLength: 2000, example: Reportado en ruta }
        cobradorId:
          type: integer
          description: Solo permitido con permiso pagos.ver_todos
          example: 1
...
    PagoPage:
      allOf:
        - $ref: '#/components/schemas/Paginated'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Pago' }
...